MapViewModuleの移行
