+++ tests/unit/components/MapView/mapWatchStoreForMoveMapCenter.spec.ts
@@ -1,5 +1,5 @@
 import map from '@/components/MapView/index.vue';
-import { mapViewGetters, mapViewMutations } from '@/store/newMapViewIndex.ts';
+import { mapViewGetters, mapViewMutations } from '@/store';
 import { testRawMapData } from '../../../resources/testRawMapData';
 import { createLocalVue, mount } from '@vue/test-utils';
 import { GeolocationWrapper } from '@/components/MapView/GeolocationWrapper.ts';
